refactor: padronizar tipo monetário e limpar código morto

+++ app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, Float, String
 from sqlalchemy.orm import relationship
 from database.database import Base
 
@@ -7,7 +7,7 @@ class User(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(256), nullable=False)
     password = Column(String(256), nullable=False)
-    budget_ceiling = Column(Integer, nullable=True)
+    budget_ceiling = Column(Float, nullable=True)
 
     incomes = relationship("Income", back_populates="user", uselist=True, cascade="all, delete-orphan")
     bills = relationship("Bill", back_populates="user", uselist=True, cascade="all, delete-orphan")
